requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
channels==4.2.0
channels-redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
kombu==5.4.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg==3.2.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
django-celery-results===2.5.1
requests===2.32.3
python-dotenv===1.0.1
openai===1.61.1